+++ js/util.js
@@ -1,0 +1,21 @@
+const getRandomNonNegativeNumber = (from, to) => {
+  if (from < 0 || to < 0 || to < from) {
+    return NaN;
+  }
+
+  if (from === to) {
+    return from;
+  }
+
+  return from + Math.floor(Math.random() * (to - from + 1));
+};
+
+const isStringLengthAllowed = (string, length) => (
+  string.length <= length
+);
+
+const getRandomArrayElement = (elements) => (
+  elements[getRandomNonNegativeNumber(0, elements.length - 1)]
+);
+
+export {getRandomNonNegativeNumber, isStringLengthAllowed, getRandomArrayElement};
